Updated Order and others 😂"

+++ src/handlers/user.ts
@@ -8,6 +8,11 @@ declare let process: {
   }
 };
 
+type authUser = {
+  firstname: string,
+  password: string
+}
+
 const store = new UserStore();
 
 export const index = async (_req: Request, res: Response) => {
@@ -30,7 +35,6 @@ export const create = async (req: Request, res: Response) => {
     const user: user = {
       firstname: req.body.firstname,
       lastname: req.body.lastname,
-      username: req.body.username,
       password: req.body.password,
     };
     const newUser = await store.create(user);
@@ -45,11 +49,15 @@ export const update = async (req: Request, res: Response) => {
   const user: user = {
     firstname: req.body.firstname,
     lastname: req.body.lastname,
-    username: req.body.username,
     password: req.body.password,
   };
 
   const id = req.params.id;
+
+  if (res.locals.user_id !== id) {
+    res.status(404).json({"error": "User does not match"});
+    return;
+  }
 
   try {
     const updatedUser = await store.update(user, id);
@@ -62,19 +70,29 @@ export const update = async (req: Request, res: Response) => {
 
 export const destroy = async (req: Request, res: Response) => {
   const id = parseInt(req.params.id);
+  
+  if (res.locals.user_id !== id) {
+    res.status(404).json({"error": "User does not match"});
+    return;
+  }
+
   const user = await store.delete(id);
   res.json(user);
 };
 
 export const authenticate =async (req: Request, res: Response) => {
   try {
-    const user = {
-      username: req.body.username,
+    const user: authUser = {
+      firstname: req.body.firstname,
       password: req.body.password,
     };
     const existingUser = await store.authenticate(user);
-    const token = jwt.sign({user: existingUser}, process.env.TOKEN_SECRET);
-    res.json(token);
+    if (existingUser === null) {
+      res.json(existingUser);
+    } else {
+      const token = jwt.sign({user: existingUser}, process.env.TOKEN_SECRET);
+      res.json(token);
+    }
   } catch (error) {
     res.status(400).json(`${error}`);
   }
