+++ src/models/order.ts
@@ -1,89 +1,79 @@
 import Client from "../database";
 
+
 export type Order = {
     id?: number,
-    status: string,
-    user_id: number
-}
-
-export type orderProducts = {
-    id?: number,
-    quantity: number,
-    orderId: string,
-    productId: string
-}
+    productId: string,
+    quantity: string,
+    user_id: string,
+    status: string
+  }
 
 export class OrderStore {
-  async index(): Promise<Order[]> {
+  async index (user_id: string): Promise<Order[]> {
     try {
       const conn = await Client.connect();
-      const sql = "SELECT * FROM orders";
-      const result = await conn.query(sql);
+      const sql = "SELECT * FROM orders WHERE user_id = $1";
+      const result = await conn.query(sql, [user_id]);
+    
       conn.release();
       return result.rows;
     } catch (error) {
-      throw new Error(`Could not get orders. Error: ${error}`);
-            
+      throw new Error("User must be authorized.");
     }
   }
-  async show(id: number): Promise<Order> {
+    
+  async show (id: number, user_id: number): Promise<Order[]> {
     try {
       const conn = await Client.connect();
-      const sql = "SELECT * FROM orders WHERE id=($1)";
-      const result = await conn.query(sql, [id]);
+      const sql = "SELECT * FROM orders WHERE id = $1 AND user_id = $2";
+      const result = await conn.query(sql, [id, user_id]);
+    
       conn.release();
-      return result.rows[0];
+      return result.rows;
     } catch (error) {
-      throw new Error(`Order with ${id} does not exist. ${error}`);
+      throw new Error("User must be authorized.");
     }
   }
-  async create(order: Order): Promise<Order> {
-    try {      
+
+  async create (o: Order) {
+    try {
       const conn = await Client.connect();
-      const sql = "INSERT INTO orders (status, user_id) VALUES ($1, $2) RETURNING *";
-      const result = await conn.query(sql, [order.status, order.user_id]);
+      const sql = "INSERT INTO orders (productId, quantity, user_id, status) VALUES ($1, $2, $3, $4) RETURNING *";
+      const result = await conn.query(sql, [o.productId, o.quantity, o.user_id, o.status]);
       conn.release();
+
       return result.rows[0];
     } catch (error) {
       throw new Error(`Could not create an order. ${error}`);
     }
   }
 
-  async addProduct(orderProduct: orderProducts): Promise<Order> {
-    try {      
+  async completeOrder (status: string, id: number, user_id: number) {
+    try {
       const conn = await Client.connect();
-      const sql = "SELECT * FROM orders WHERE id = ($1)";
-      const result = await conn.query(sql, [orderProduct.orderId]);
-      const order = result.rows[0];
+      const sql = "UPDATE orders SET status = $1 WHERE id = $2 AND user_id = $3 RETURNING *";
+      const result = await conn.query(sql, [status, id, user_id]);
+      conn.release();
 
-      if (order.status !== "active" ) {
-        throw new Error("Cannot add products to this order, order is not active.");
-      }
-      conn.release();
-    } catch (error) {
-      throw new Error(`${error}`);
-    }
-    try {      
-      const conn = await Client.connect();
-      const sql = "INSERT INTO order_products (quantity, order_id, product_id) VALUES ($1, $2, $3) RETURNING *";
-      const result = await conn.query(sql, [orderProduct.quantity, orderProduct.orderId, orderProduct.productId]);
-      conn.release();
       return result.rows[0];
     } catch (error) {
-      throw new Error(`Could not add product(s) to order. ${error}`);
+      throw new Error(`Could not complete order. ${error}`);
+            
     }
   }
 
-  async delete(id: number): Promise<string> {
+  async delete(id: number, user_id: number): Promise<string> {
     try {
       const conn = await Client.connect();
-      const sql = "DELETE FROM orders WHERE id=($1)";
-      await conn.query(sql, [id]);
+      const sql = "DELETE FROM orders WHERE id=$1 AND user_id = $2";
+      await conn.query(sql, [id, user_id]);
       conn.release();
+          
       return "Deleted successfully!";
     } catch (error) {
       throw new Error(`Could not delete order ${id}. ${error}`);
-            
+                
     }
   }
 }
